requirements.txt
Django==5.0.3
pillow==10.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
XlsxWriter==3.2.0
